+++ lib/github_changelog_generator/generator/issue.rb
@@ -1,0 +1,9 @@
+module GitHubChangelogGenerator
+  class Issue
+    attr_accessor :name
+
+    def initialize(opts = {})
+      @name = opts[:name]
+    end
+  end
+end
